requirements.txt
aiogram==2.15
sqlalchemy==1.4.25
python-dotenv==0.17.0
psycopg2==2.9.1
loguru==0.5.3